Streamline mail components filenames (#920)
