+++ libs/mails/src/index.ts
@@ -1,6 +1,4 @@
-export * from './lib/mailProvider'
-export * from './lib/mailContext'
-export * from './lib/MailchimpMailProvider'
-export * from './lib/MailgunMailProvider'
+export * from './lib/mail-provider'
 export * from './lib/mail.controller'
+export * from './lib/mail-context'
 export * from './lib/mails.module'
