+++ libs/mails/src/lib/mail-provider/index.ts
@@ -1,0 +1,4 @@
+export * from './mail-provider.interface'
+export * from './base-mail-provider'
+export * from './mailgun-mail-provider'
+export * from './mailchimp-mail-provider'
