+++ libs/mails/src/lib/mail-provider/mail-provider.interface.ts
@@ -1,5 +1,4 @@
 import {MailLogState} from '@prisma/client'
-import bodyParser from 'body-parser'
 import {NextHandleFunction} from 'connect'
 import express from 'express'
 
@@ -70,34 +69,4 @@ export interface MailProvider {
   getTemplates(): Promise<MailProviderTemplate[]>
 
   getTemplateUrl(template: WithExternalId): string
-}
-
-export interface MailProviderProps {
-  id: string
-  name: string
-  fromAddress: string
-  incomingRequestHandler?: NextHandleFunction
-}
-
-export abstract class BaseMailProvider implements MailProvider {
-  readonly id: string
-  readonly name: string
-  readonly fromAddress: string
-
-  readonly incomingRequestHandler: NextHandleFunction
-
-  protected constructor(props: MailProviderProps) {
-    this.id = props.id
-    this.name = props.name
-    this.fromAddress = props.fromAddress
-    this.incomingRequestHandler = props.incomingRequestHandler ?? bodyParser.json()
-  }
-
-  abstract webhookForSendMail(props: WebhookForSendMailProps): Promise<MailLogStatus[]>
-
-  abstract sendMail(props: SendMailProps): Promise<void>
-
-  abstract getTemplates(): Promise<MailProviderTemplate[]>
-
-  abstract getTemplateUrl(template: WithExternalId): string
 }
